Move layermenu from top bar to left panel
