+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuItemMobile/index.test.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
 import MenuItemMobile from '.';
-import { store } from '../../../context/store';
-import { LayerKey, MenuItemMobileType } from '../../../config/types';
+import { store } from '../../../../../context/store';
+import { LayerKey, MenuItemMobileType } from '../../../../../config/types';
 
 jest.mock('../MenuSwitch', () => 'mock-MenuSwitch');
