+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuSwitch/index.test.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
 import MenuSwitch from '.';
-import { store } from '../../../context/store';
-import { LayerKey, LayersCategoryType } from '../../../config/types';
+import { store } from '../../../../../context/store';
+import { LayerKey, LayersCategoryType } from '../../../../../config/types';
 
 jest.mock('./SwitchItem', () => 'mock-SwitchItem');
