+++ frontend/src/components/MapView/LeftPanel/layersPanel/index.tsx
@@ -1,0 +1,19 @@
+import { Box } from '@material-ui/core';
+import React from 'react';
+// import MenuItemMobile from './MenuItemMobile';
+// import { menuList } from './utils';
+
+function LayersPanel() {
+  // const menuMobile = menuList.map(({ title, ...category }) => (
+  //     <MenuItemMobile
+  //       expanded={expanded}
+  //       selectAccordion={selectAccordion}
+  //       key={title}
+  //       title={title}
+  //       {...category}
+  //     />
+  //   ));
+  return <Box>TEST</Box>;
+}
+
+export default LayersPanel;
